feat: bulk & synchronous APNS pushing
APNS.push/2 now takes either a single notification or list as its first
argument. If an `:on_response` key is defined in the opts, the push
happens asynchronously, else the normal `{:ok, notif}` or `{:error,
reason, notif}` tuple is returned.

For async bulk pushes, on_response is triggered for each individual
notification. For synchronous, they are grouped into a map like so:
```
%{
  ok: [notif_1, notif_2],
  error: %{
    bad_device_token: [notif_3],
    bad_topic: [notif_4]
    # etc
  }
}
```