Clarification to keep key conversion from failing
If the PEM pass phrase is less than 4 characters it fails and there are no good error messages, google searches, or stack overflow posts to help you understand why it's failing. I added text to warn the reader of this.

I also changed the wording from "password" to "pass phrase" because that is the word used by the openssl script in the terminal. This help distinguish from the password that can, but not must be set when exporting the certificate or passkey.