chore: update Joken (#150)
Joken 2.1 was released and it is a minor update only because we removed the unnecessary application startup.

https://github.com/joken-elixir/joken/releases/tag/v2.1.0

It also clean ups some warnings so I thought it would be good to update here too to avoid having to override it. I've also used the minor version only here to avoid locking possible patch updates.